+++ tpls/gtest/gtest/gtest-all.cc
@@ -7565,7 +7565,7 @@ static pid_t ExecDeathTestSpawnChild(char* const* argv, int close_fd) {
         static_cast<char*>(stack) +
             (stack_grows_down ? stack_size - kMaxStackAlignment : 0);
     GTEST_DEATH_TEST_CHECK_(stack_size > kMaxStackAlignment &&
-        reinterpret_cast<intptr_t>(stack_top) % kMaxStackAlignment == 0);
+        reinterpret_cast<int64_t>(stack_top) % kMaxStackAlignment == 0);
 
     child_pid = clone(&ExecDeathTestChildMain, stack_top, SIGCHLD, &args);
